+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/h2/database/io/H2ExtrasLeafIO.java
@@ -84,7 +84,7 @@ public class H2ExtrasLeafIO extends BPlusLeafIO<SearchRow> {
         assert srcPageAddr == dstItemSize || dstItemSize == 0;
 
         if (dstItemSize == 0)
-            writeMetaHeader(dstPageAddr, srcItemSize);
+            writeMetaHeader(dstPageAddr, (short)srcItemSize);
 
         int srcOff = srcIo.offset(srcPageAddr, srcIdx);
         byte[] payload = PageUtils.getBytes(srcPageAddr, srcOff, srcItemSize);
@@ -117,21 +117,20 @@ public class H2ExtrasLeafIO extends BPlusLeafIO<SearchRow> {
 
     /** {@inheritDoc} */
     @Override public void writeMetaHeader(long pageAddr, Object obj) {
-        short size = obj == null ? 0 : (Short)obj;
+        short size = obj == null ? 0 : (short)obj;
         PageUtils.putShort(pageAddr, META_HEADER_OFFSET, size);
     }
 
-
     /** {@inheritDoc} */
     @Override public int itemSize(long pageAddr) {
-        int itemSize = PageUtils.getInt(pageAddr, META_HEADER_OFFSET);
+        int itemSize = PageUtils.getShort(pageAddr, META_HEADER_OFFSET);
 
         if (itemSize != 0)
             return itemSize;
         H2TreeIndex currentIndex = H2TreeIndex.getCurrentIndex();
         assert currentIndex != null;
 
-        writeMetaHeader(pageAddr, currentIndex.itemSize());
+        writeMetaHeader(pageAddr, (short)currentIndex.itemSize());
         return currentIndex.itemSize();
     }
 }
